+++ cmd/bdjuno/main.go
@@ -17,7 +17,7 @@ import (
 	"github.com/forbole/bdjuno/v3/modules"
 
 	cmdxapp "github.com/comdex-official/comdex/app"
-	gaiaapp "github.com/cosmos/gaia/v6/app"
+	gaiaapp "github.com/cosmos/gaia/v7/app"
 )
 
 func main() {