+++ cmd/parse/parse.go
@@ -12,6 +12,7 @@ import (
 	parsefeegrant "github.com/forbole/bdjuno/v3/cmd/parse/feegrant"
 	parsegov "github.com/forbole/bdjuno/v3/cmd/parse/gov"
 	parsestaking "github.com/forbole/bdjuno/v3/cmd/parse/staking"
+	parsewasm "github.com/forbole/bdjuno/v3/cmd/parse/wasm"
 	parsetransaction "github.com/forbole/juno/v3/cmd/parse/transactions"
 )
 
@@ -31,6 +32,7 @@ func NewParseCmd(parseCfg *parse.Config) *cobra.Command {
 		parsegov.NewGovCmd(parseCfg),
 		parsestaking.NewStakingCmd(parseCfg),
 		parsetransaction.NewTransactionsCmd(parseCfg),
+		parsewasm.NewWasmCmd(parseCfg),
 	)
 
 	return cmd